+++ lib/markdown2.py
@@ -1988,14 +1988,48 @@ class Markdown:
         self._code_table[text] = hashed
         return hashed
 
-    _strong_re = re.compile(r"(\*\*|__)(?=\S)(.+?[*_]?)(?<=\S)\1", re.S)
+    _strong_re = re.compile(r'''
+        (?:_{1,}|\*{1,})?  # ignore any leading em chars because we want to wrap `<strong>` as tightly around the text as possible
+                           # eg: `***abc***` -> `*<strong>abc</strong>*` instead of `<strong>*abc*</strong>`
+                           # Makes subsequent <em> processing easier
+        (\*\*|__)(?=\S)    # strong syntax - must be followed by a non whitespace char
+        (.+?)              # the strong text itself
+        (?<=\S)\1          # closing syntax - must be preceeded by non whitespace char
+        ''',
+        re.S | re.X
+    )
     _em_re = re.compile(r"(\*|_)(?=\S)(.*?\S)\1", re.S)
 
     @mark_stage(Stage.ITALIC_AND_BOLD)
     def _do_italics_and_bold(self, text: str) -> str:
+        def sub(match: re.Match):
+            '''
+            regex sub function that checks that the match isn't matching across spans.
+            The span shouldn't be across a closing or opening HTML tag, although spans within
+            the span is acceptable.
+            '''
+            contents: str = match.group(2)
+            # the strong re also checks for leading em chars, so the match may cover some additional text
+            prefix = match.string[match.start(): match.regs[1][0]]
+            # look for all possible span HTML tags
+            for tag in re.findall(rf'</?({self._span_tags})', contents):
+                # if it's unbalanced then that violates the rules
+                if not self._tag_is_closed(tag, contents):
+                    return prefix + match.group(1) + contents + match.group(1)
+
+                # if it is balanced, but the closing tag is before the opening then
+                # the text probably looks like `_</strong>abcdef<strong>_`, which is across 2 spans
+                close_index = contents.find(f'</{tag}')
+                open_index = contents.find(f'<{tag}')
+                if close_index != -1 and close_index < open_index:
+                    return prefix + match.group(1) + contents + match.group(1)
+
+            syntax = 'strong' if len(match.group(1)) == 2 else 'em'
+            return f'{prefix}<{syntax}>{contents}</{syntax}>'
+
         # <strong> must go first:
-        text = self._strong_re.sub(r"<strong>\2</strong>", text)
-        text = self._em_re.sub(r"<em>\2</em>", text)
+        text = self._strong_re.sub(sub, text)
+        text = self._em_re.sub(sub, text)
         return text
 
     _block_quote_base = r'''
@@ -3320,7 +3354,7 @@ class MiddleWordEm(ItalicAndBoldProcessor):
         self.middle_word_em_re = re.compile(
             r'''
             (?<!^)         # To be middle of a word, it cannot be at the start of the input
-            (?<![*_\s])    # cannot be preceeded by em character or whitespace (must be in middle of word)
+            (?<![*_\W])    # cannot be preceeded by em char or non word char (must be in middle of word)
             ([*_])         # em char
             (?=\S)         # must be followed by non-whitespace char
             (?![*_]|$|\W)  # cannot be followed by another em char, EOF or a non-word char